Linking, debug and panel ready for prod

+++ libs/program_gui/basic_elements/GUI_functions/translators/color_translator.h
@@ -21,4 +21,4 @@
  *
  * @return SDL_Color Resulting color with RGBA components.
  */
-SDL_Color hex_to_sdl_color(const std::string& hex, unsigned int opacity);
+SDL_Color hex_to_sdl_color(const std::string& hex, Uint8 opacity);
